Update RPATH handling (#126)
This changes the default behavior of TriBITS projects to set RPATH on install
so that everything works right out of the install directory without having to
set any extra env vars.

Several things done as part of this:

* Added SimpleTpl to help teste RPATH.  (This will also come in handy for
  doing more TPL testing.)

* Add installable exec to SimpleCxx package to help test RPATH.

* Added documentation for users and developers (see documenation files).

* Added detailed automated tests to test the different use cases described in
  the documentation.

I manually tested this locally on Linux, Mac OSX (gaia) and on Travis CI (by
pushing the branch fix-rpath-126 to the main TriBITS GitHub repo).  I think
this is pretty solid.  See TribitsPub/TriBITS#126 for more motivation and
details.