WIP: Make RPATH mostly pass on OSX (#126)
By configuring the SimpleTpl install with:

  -DCMAKE_INSTALL_RPATH_USE_LINK_PATH=TRUE
  -DCMAKE_MACOSX_RPATH=TRUE

now RPATH is handled in the downstream TriBITS installed executables.

However, one test still fails which is where you set the final RPATH with
CMAKE_INSTALL_RPATH.  This should work but it does not work for some reason.

I don't know that this is such an important use case but I want to check to
see if this can be fixed before I disable this test on this platform and
document that this use case does not work on OSX.