Disable test TribitsExampleProject_WrapExternal on OSX
Tim Fuller researched this this is likely coming from
tribits/core/package_arch/TribitsFindMostRecentFileTimestamp.cmake.  The
following two commands seem problematic:

    COMMAND ls --full-time "${CURRENT_FILEPATH}"
    COMMAND find . -type f -printf "%T@ %p\n"

The “--full-time” and “–printf” options are not valid for BSD ls and find.  If
I use the gnu ls and gnu find (“gls” and “gfind” installed through Homebrew)
it still fails.  I suspect the output from ls --full-time is different than on
linux since the test fails from differing time stamps.

This feature is not a critical cross-platform feature and even for CASL VERA
Bison/MOOSE, we can disable this feature on a given configure.  It is really
just there to try to give the hybrid CMake project some automatic dependency
checking like you get for a CMake project.  If this runs only on Linux, that
is fine by me.

Therefore, I am disabling this test on OSX for now.